requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
streamlit>=1.28.0
PyYAML>=6.0
tqdm>=4.66.0
SoccerNet>=0.1.50
trackeval>=1.0.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
pytest>=7.4.0
roboflow>=1.1.20
